Redesign auto update events
