+++ src/public/about-dialog.html
@@ -42,7 +42,7 @@
 					document.querySelector('.check-for-updates').setAttribute('disabled', 'disabled');
 					document.querySelector('.check-for-updates').classList.add('hidden');
 					document.querySelector('.checking-for-updates').classList.remove('hidden');
-					ipcRenderer.send('check-for-updates');
+					ipcRenderer.send('check-for-updates', { forced: true });
 				}, false);
 
 				const resetUpdatesSection = () => {
